refactor demo responses into separate file
